backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
firebase-admin
